feat: allow dynamic adjustment of layout ratio via `rt.mgr.ratio`
