+++ yazi-binding/Cargo.toml
@@ -18,11 +18,12 @@ yazi-shared  = { path = "../yazi-shared", version = "25.6.11" }
 
 # External dependencies
 ansi-to-tui   = { workspace = true }
+foldhash      = { workspace = true }
 mlua          = { workspace = true }
 paste         = { workspace = true }
 ratatui       = { workspace = true }
 serde_json    = { workspace = true }
-tokio		      = { workspace = true }
+tokio  	      = { workspace = true }
 unicode-width = { workspace = true }
 
 # Logging
