+++ yazi-binding/src/elements/text.rs
@@ -32,12 +32,7 @@ impl Text {
 		text.into_lua(lua)
 	}
 
-	pub(super) fn render(
-		self,
-		buf: &mut ratatui::buffer::Buffer,
-		trans: impl Fn(yazi_config::popup::Position) -> ratatui::layout::Rect,
-	) {
-		let rect = self.area.transform(trans);
+	pub(super) fn render(self, rect: ratatui::layout::Rect, buf: &mut ratatui::buffer::Buffer) {
 		if self.wrap.is_none() && self.scroll == Default::default() {
 			self.inner.render(rect, buf);
 		} else {
