+++ yazi-config/src/mgr/mgr.rs
@@ -2,26 +2,27 @@ use anyhow::{Result, bail};
 use serde::Deserialize;
 use yazi_codegen::DeserializeOver2;
 use yazi_fs::{CWD, SortBy};
+use yazi_shared::SyncCell;
 
 use super::{MgrRatio, MouseEvents};
 
 #[derive(Debug, Deserialize, DeserializeOver2)]
 pub struct Mgr {
-	pub ratio: MgrRatio,
+	pub ratio: SyncCell<MgrRatio>,
 
 	// Sorting
-	pub sort_by:        SortBy,
-	pub sort_sensitive: bool,
-	pub sort_reverse:   bool,
-	pub sort_dir_first: bool,
-	pub sort_translit:  bool,
+	pub sort_by:        SyncCell<SortBy>,
+	pub sort_sensitive: SyncCell<bool>,
+	pub sort_reverse:   SyncCell<bool>,
+	pub sort_dir_first: SyncCell<bool>,
+	pub sort_translit:  SyncCell<bool>,
 
 	// Display
 	pub linemode:     String,
-	pub show_hidden:  bool,
-	pub show_symlink: bool,
-	pub scrolloff:    u8,
-	pub mouse_events: MouseEvents,
+	pub show_hidden:  SyncCell<bool>,
+	pub show_symlink: SyncCell<bool>,
+	pub scrolloff:    SyncCell<u8>,
+	pub mouse_events: SyncCell<MouseEvents>,
 	pub title_format: String,
 }
