+++ yazi-config/src/mgr/ratio.rs
@@ -2,7 +2,7 @@ use anyhow::bail;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(try_from = "Vec<u16>")]
+#[serde(try_from = "[u16; 3]")]
 pub struct MgrRatio {
 	pub parent:  u16,
 	pub current: u16,
@@ -10,10 +10,10 @@ pub struct MgrRatio {
 	pub all:     u16,
 }
 
-impl TryFrom<Vec<u16>> for MgrRatio {
+impl TryFrom<[u16; 3]> for MgrRatio {
 	type Error = anyhow::Error;
 
-	fn try_from(ratio: Vec<u16>) -> Result<Self, Self::Error> {
+	fn try_from(ratio: [u16; 3]) -> Result<Self, Self::Error> {
 		if ratio.len() != 3 {
 			bail!("invalid layout ratio: {:?}", ratio);
 		}
