+++ yazi-core/src/tab/preference.rs
@@ -19,15 +19,15 @@ impl Default for Preference {
 	fn default() -> Self {
 		Self {
 			// Sorting
-			sort_by:        YAZI.mgr.sort_by,
-			sort_sensitive: YAZI.mgr.sort_sensitive,
-			sort_reverse:   YAZI.mgr.sort_reverse,
-			sort_dir_first: YAZI.mgr.sort_dir_first,
-			sort_translit:  YAZI.mgr.sort_translit,
+			sort_by:        YAZI.mgr.sort_by.get(),
+			sort_sensitive: YAZI.mgr.sort_sensitive.get(),
+			sort_reverse:   YAZI.mgr.sort_reverse.get(),
+			sort_dir_first: YAZI.mgr.sort_dir_first.get(),
+			sort_translit:  YAZI.mgr.sort_translit.get(),
 
 			// Display
 			linemode:    YAZI.mgr.linemode.to_owned(),
-			show_hidden: YAZI.mgr.show_hidden,
+			show_hidden: YAZI.mgr.show_hidden.get(),
 		}
 	}
 }
