--- yazi-plugin/src/config/mod.rs
@@ -1,1 +1,0 @@
-yazi_macro::mod_flat!(plugin runtime term theme);
