+++ yazi-plugin/src/fs/fs.rs
@@ -7,7 +7,7 @@ use yazi_fs::{mounts::PARTITIONS, remove_dir_clean};
 use crate::bindings::SizeCalculator;
 
 pub fn compose(lua: &Lua) -> mlua::Result<Value> {
-	Composer::make(lua, |lua, key| {
+	fn get(lua: &Lua, key: &[u8]) -> mlua::Result<Value> {
 		match key {
 			b"op" => op(lua)?,
 			b"cwd" => cwd(lua)?,
@@ -23,7 +23,11 @@ pub fn compose(lua: &Lua) -> mlua::Result<Value> {
 			_ => return Ok(Value::Nil),
 		}
 		.into_lua(lua)
-	})
+	}
+
+	fn set(_: &Lua, _: &[u8], value: Value) -> mlua::Result<Value> { Ok(value) }
+
+	Composer::make(lua, get, set)
 }
 
 fn op(lua: &Lua) -> mlua::Result<Function> {
