+++ yazi-plugin/src/isolate/isolate.rs
@@ -11,8 +11,8 @@ pub fn slim_lua(name: &str) -> mlua::Result<Lua> {
 	globals.raw_set("ui", yazi_binding::elements::compose(&lua)?)?;
 	globals.raw_set("ya", crate::utils::compose(&lua, true)?)?;
 	globals.raw_set("fs", crate::fs::compose(&lua)?)?;
-	globals.raw_set("rt", crate::config::Runtime::compose(&lua)?)?;
-	globals.raw_set("th", crate::config::Theme::compose(&lua)?)?;
+	globals.raw_set("rt", crate::runtime::compose(&lua)?)?;
+	globals.raw_set("th", crate::theme::compose(&lua)?)?;
 
 	yazi_binding::Cha::install(&lua)?;
 	yazi_binding::File::install(&lua)?;
