+++ yazi-plugin/src/lib.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::if_same_then_else, clippy::unit_arg)]
 
-yazi_macro::mod_pub!(bindings config external fs isolate loader process pubsub utils);
+yazi_macro::mod_pub!(bindings external fs isolate loader process pubsub runtime theme utils);
 
 yazi_macro::mod_flat!(lua twox);
