+++ yazi-plugin/src/runtime/mod.rs
@@ -1,0 +1,1 @@
+yazi_macro::mod_flat!(plugin runtime term);
