+++ yazi-plugin/src/theme/mod.rs
@@ -1,0 +1,1 @@
+yazi_macro::mod_flat!(theme);
